requirements.txt
anthropic>=0.92.0
openai>=1.30.0
gspread>=6.0.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
requests>=2.31.0
Pillow>=10.0.0
python-dotenv>=1.0.0
jinja2>=3.1.0
beautifulsoup4>=4.12.0
slack-sdk>=3.0.0
filelock>=3.28.0
httpx>=0.28.0
rapidfuzz>=3.0.0
openpyxl>=3.1.0
facebook-business>=19.0
google-ads>=25.0
cairosvg>=2.7
